+++ apps/liuli-cli/src/__tests__/addHusky.test.ts
@@ -1,7 +1,0 @@
-import { addHusky } from '../addHusky'
-
-describe('测试 addHusky', () => {
-  it('测试添加依赖', () => {
-    addHusky()
-  })
-})