+++ apps/liuli-cli/src/clean.ts
@@ -1,9 +1,0 @@
-import { pathExists, remove } from 'fs-extra'
-import * as path from 'path'
-
-export async function clean() {
-  const distPath = path.resolve('./dist')
-  if (await pathExists(distPath)) {
-    await remove(distPath)
-  }
-}