+++ apps/liuli-cli/src/commands/init/index.ts
@@ -4,23 +4,23 @@ import {
   GenerateProgram,
   TemplateTypeEnum,
 } from './GenerateProgram'
-import path from 'path'
 
 const initProgram = new GenerateProgram()
 
-const templateOption = new Option('--template <template>', '模板类型').choices([
+const templateOption = new Option('--template [template]', '模板类型').choices([
   TemplateTypeEnum.Lib,
   TemplateTypeEnum.Cli,
 ])
 templateOption.required = true
 
 export const initCommand = new Command()
-  .command('generate <dest>')
+  .command('generate [dest]')
   .description('生成一些初始项目')
   .addOption(templateOption)
   .action(async (dest, options: Pick<GenerateConfig, 'template'>) => {
     await initProgram.generate({
       ...options,
-      dest: path.resolve(dest),
+      dest: dest,
+      initSync: true,
     })
   })