+++ apps/liuli-cli/src/index.ts
@@ -1,2 +1,4 @@
 export * from './commands/build/BuildProgram'
+export * from './commands/init/GenerateProgram'
+export * from './commands/sync/SyncProgram'
 export * from './utils'