+++ apps/liuli-cli/src/utils/__tests__/arrayToMap.test.ts
@@ -1,0 +1,37 @@
+import { arrayToMap } from '../arrayToMap'
+
+/**
+ * @test {arrayToMap}
+ */
+describe('test arrayToMap', () => {
+  it('arrayToMap [1, 2, 3] to equals Map({1 => 1, 2 => 2, 3 => 3})', () => {
+    const arr = [1, 2, 3]
+    expect(arrayToMap(arr, (i: any) => i)).toEqual(
+      new Map().set(1, 1).set(2, 2).set(3, 3),
+    )
+  })
+  it('use vFn', () => {
+    const arr = [1, 2, 3]
+    expect(
+      arrayToMap(
+        arr,
+        (v) => v,
+        (i) => i * 2 + '',
+      ),
+    ).toEqual(new Map().set(1, '2').set(2, '4').set(3, '6'))
+  })
+  it('use string k/v field', () => {
+    class User {
+      constructor(public id: number, public name: string) {}
+    }
+
+    const arr = [new User(1, 'rx'), new User(2, '琉璃'), new User(3, '楚轩')]
+    expect(
+      arrayToMap(
+        arr,
+        (item) => item.id,
+        (item) => item.name,
+      ),
+    ).toEqual(new Map().set(1, 'rx').set(2, '琉璃').set(3, '楚轩'))
+  })
+})