+++ apps/liuli-cli/src/utils/__tests__/test.config.ts
@@ -1,6 +1,0 @@
-import path from 'path'
-import { RollupOptions } from 'rollup'
-
-export default {
-  input: path.resolve(__dirname, 'src/index.ts'),
-} as RollupOptions