+++ apps/liuli-cli/src/utils/__tests__/ts-node-run.ts
@@ -1,4 +1,0 @@
-import path from 'path'
-import { loadTypeScriptConfig } from '../loadTypeScriptConfig'
-
-console.log(loadTypeScriptConfig(path.resolve(__dirname, 'test.config.ts')))