+++ apps/liuli-cli/src/utils/appendScript.ts
@@ -1,0 +1,17 @@
+/**
+ * 在之前的脚本中追加新的脚本
+ * @param oldScript
+ * @param newScript
+ */
+export function appendScript(
+  oldScript: string | undefined,
+  newScript: string,
+): string {
+  if (!oldScript) {
+    return newScript
+  }
+  if (oldScript.includes(newScript)) {
+    return oldScript
+  }
+  return oldScript + ' && ' + newScript
+}